+++ actionpack/CHANGELOG.md
@@ -1,2 +1,5 @@
+*   Fix reloader deadlock in ActionController Live by adding timeout loop to ActionDispatch Response#await_commit
+
+    *Nate Matykiewicz, Ben Sheldon*
 
 Please check [8-0-stable](https://github.com/rails/rails/blob/8-0-stable/actionpack/CHANGELOG.md) for previous changes.